refactor: reference vite logo with img tag, adding style tag inside svg
