+++ packages/create-vite/template-vanilla/src/main.js
@@ -1,6 +1,6 @@
 import './style.css'
 import javascriptLogo from './assets/javascript.svg'
-import viteLogo from './assets/vite.svg?raw'
+import viteLogo from './assets/vite.svg'
 import heroImg from './assets/hero.png'
 import { setupCounter } from './counter.js'
 
@@ -9,7 +9,7 @@ document.querySelector('#app').innerHTML = `
   <div class="hero">
     <img src="${heroImg}" class="base" width="170" height="179">
     <img src="${javascriptLogo}" class="framework" alt="JavaScript logo"/>
-    <span class="vite" aria-label="Vite logo">${viteLogo}</span>
+    <img src=${viteLogo} class="vite" alt="Vite logo" />
   </div>
   <div>
     <h1>Get started</h1>
@@ -28,7 +28,7 @@ document.querySelector('#app').innerHTML = `
     <ul>
       <li>
         <a href="https://vite.dev/" target="_blank">
-          <span class="logo">${viteLogo}</span>
+          <img class="logo" src=${viteLogo} alt="" />
           Explore Vite
         </a>
       </li>
