+++ packages/create-vite/template-vanilla/src/style.css
@@ -8,7 +8,6 @@
   --accent-bg: rgba(170, 59, 255, 0.1);
   --accent-border: rgba(170, 59, 255, 0.5);
   --social-bg: rgba(244, 243, 236, 0.5);
-  --vite-logo: #000;
   --shadow:
     rgba(0, 0, 0, 0.1) 0 10px 15px -3px, rgba(0, 0, 0, 0.05) 0 4px 6px -2px;
 
@@ -42,7 +41,6 @@
     --accent-bg: rgba(192, 132, 252, 0.15);
     --accent-border: rgba(192, 132, 252, 0.5);
     --social-bg: rgba(47, 48, 58, 0.5);
-    --vite-logo: white;
     --shadow:
       rgba(0, 0, 0, 0.4) 0 10px 15px -3px, rgba(0, 0, 0, 0.25) 0 4px 6px -2px;
   }
@@ -151,16 +149,10 @@ code {
   .vite {
     z-index: 0;
     top: 107px;
-    display: flex;
-    justify-content: center;
-    color: var(--vite-logo);
+    height: 26px;
+    width: auto;
     transform: perspective(2000px) rotateZ(300deg) rotateX(40deg) rotateY(39deg)
       scale(0.8);
-
-    svg {
-      height: 26px;
-      width: auto;
-    }
   }
 }
 
@@ -232,11 +224,6 @@ code {
   margin: 32px 0 0;
 
   .logo {
-    color: var(--vite-logo);
-    svg {
-      height: 100%;
-      width: auto;
-    }
     height: 18px;
   }
